Move builtin compile task types to types module.
