+++ packages/hardhat-core/src/types/builtin-tasks/compile.ts
@@ -1,4 +1,4 @@
-import { SolcConfig } from "../types/config";
+import { SolcConfig } from "../config";
 
 /**
  * A Solidity file.
