+++ packages/hardhat-core/src/types/builtin-tasks/index.ts
@@ -1,0 +1,1 @@
+export * from "./compile";
